chore: re-organize code
